+++ service/provider/src/main/java/gov/va/vro/service/camel/ClaimProcessorRoute.java
@@ -1,9 +1,8 @@
-package gov.va.vro.routes;
+package gov.va.vro.service.camel;
 
 import gov.va.starter.example.service.spi.claimsubmission.model.ClaimSubmission;
-import gov.va.vro.CamelDtoConverter;
 import gov.va.vro.model.Payload;
-import gov.va.vro.services.ClaimProcessorA;
+import gov.va.vro.service.processors.ClaimProcessorA;
 import org.apache.camel.ExchangeProperties;
 import org.apache.camel.builder.RouteBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
